# pendaftar/transaksi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "API Documentation",
        "contact": {
            "email": "arofanurhadi@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/license/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Documentation"
        }
    ],
    "paths": {
        "/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get All Books",
                "description": "Retrieve all Books data",
                "operationId": "listBooks",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Berhasil Mengambil Data Books",
                                    "data": [
                                        {
                                            "id": "1",
                                            "title": "Bersamamu",
                                            "author": "Dina",
                                            "years": "2024"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Books Get By ID",
                "operationId": "listBooksById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Books",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Berhasil mengambil Data Books",
                                    "data": {
                                        "id": "1",
                                        "title": "Bersamamu",
                                        "author": "Dina",
                                        "years": "2024"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/books/create": {
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Books Insert",
                "description": "Tambah data Books",
                "operationId": "insertBooks",
                "requestBody": {
                    "description": "Data Books yang ingin ditambahkan",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "id": "1",
                                "title": "Bersamamu",
                                "author": "Dina",
                                "years": "2024"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Data berhasil disimpan"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/books/update": {
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Books Update",
                "description": "Update data Books",
                "operationId": "updateBooks",
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "id": "1",
                                "title": "Bersamamu",
                                "author": "Dina",
                                "years": "2024"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Data berhasil diubah"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/books/delete": {
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Books Delete",
                "description": "Hapus data Books",
                "operationId": "deleteBooks",
                "requestBody": {
                    "description": "ID Books yang ingin dihapus",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "id": "1"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Data berhasil dihapus"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/chapters/update": {
            "put": {
                "tags": [
                    "Chapters"
                ],
                "summary": "Chapters - Update",
                "description": "Update data Chapters",
                "operationId": "updateChapters",
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "id": "9",
                                "book_id": "2",
                                "title": "Bab 11:Semangat"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Data berhasil diubah"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token ",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Books",
            "description": "Books"
        },
        {
            "name": "Books",
            "description": "Books"
        },
        {
            "name": "Books",
            "description": "Books"
        },
        {
            "name": "Books",
            "description": "Books"
        },
        {
            "name": "Books",
            "description": "Books"
        },
        {
            "name": "Chapters",
            "description": "Chapters"
        }
    ]
}